requirements.txt
Django>=3.2,<4.0
djangorestframework
gunicorn
psycopg2-binary
django-ratelimit
ratelimit
django-use-email-as-username